backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.7
redis==5.0.1
django-redis==5.4.0
Pillow==10.0.1
django-filter==23.3
djangorestframework-simplejwt==5.3.0
celery==5.3.4
gunicorn==21.2.0
